Refactor proxy tests and remove unused methods
