+++ spec/spec_helper.rb
@@ -14,25 +14,10 @@ require_relative "../lib/twingly/http"
 
 require_relative "spec_help/env_helper"
 require_relative "spec_help/fixture"
+require_relative "spec_help/http_helpers"
 require_relative "spec_help/test_logger"
 require_relative "spec_help/toxiproxy_config"
-require_relative "spec_help/http_test_server"
 require_relative "spec_help/port_prober"
-
-module HttpHelpers
-  def with_real_http_connections
-    original_cassette = VCR.current_cassette
-    VCR.eject_cassette if original_cassette
-    VCR.turn_off!
-    WebMock.allow_net_connect!
-
-    yield
-  ensure
-    WebMock.disable_net_connect!
-    VCR.turn_on!
-    VCR.insert_cassette(original_cassette.name) if original_cassette
-  end
-end
 
 # Start with a clean slate, destroy all proxies if any
 Toxiproxy.all.destroy
